requirements.txt
applescript==2021.2.9
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
googleapis-common-protos==1.63.0
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
iniconfig==2.0.0
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
protobuf==4.25.3
pyperclip==1.8.2
PySocks==1.7.1
pytest==8.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
Random-Word==1.0.11
requests==2.31.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
viam-sdk==0.16.1
wsproto==1.2.0
